refactor(test): split large FVT step definition files
Keep godog registrations and shared types in place while moving MCP
and Kubernetes step bodies into focused same-package files so the
largest FVT harnesses are easier to navigate.
