+++ tests/kubernetes/features/service_account_steps_test.go
@@ -1,0 +1,102 @@
+package features
+
+import (
+	"fmt"
+	"os"
+	"strings"
+
+	"github.com/cucumber/godog"
+)
+
+// ============================================================================
+// Service Account & Environment Steps
+// ============================================================================
+
+func (tc *testContext) mlflowIsConfigured() error {
+	if os.Getenv("MLFLOW_TRACKING_URI") == "" {
+		return godog.ErrSkip
+	}
+	return nil
+}
+
+func (tc *testContext) environmentVariableIsSet(name, value string) error {
+	if os.Getenv("K8S_TEST_DEBUG") == "true" {
+		fmt.Printf("[DEBUG] assuming service env %s=%s\n", name, value)
+	}
+	return nil
+}
+
+func (tc *testContext) containerCommandShouldBeValidArray() error {
+	if tc.currentJob == nil {
+		return fmt.Errorf("no current Job")
+	}
+
+	if len(tc.currentJob.Spec.Template.Spec.Containers) == 0 {
+		return fmt.Errorf("Job %s has no containers", tc.currentJob.Name)
+	}
+
+	container := tc.currentJob.Spec.Template.Spec.Containers[0]
+	if len(container.Command) == 0 {
+		return fmt.Errorf("Container %s has no command", container.Name)
+	}
+
+	return nil
+}
+
+func (tc *testContext) containerCommandShouldNotContainEmptyStrings() error {
+	if tc.currentJob == nil {
+		return fmt.Errorf("no current Job")
+	}
+
+	if len(tc.currentJob.Spec.Template.Spec.Containers) == 0 {
+		return fmt.Errorf("Job %s has no containers", tc.currentJob.Name)
+	}
+
+	container := tc.currentJob.Spec.Template.Spec.Containers[0]
+	for _, cmd := range container.Command {
+		if cmd == "" {
+			return fmt.Errorf("Container %s command contains empty string", container.Name)
+		}
+	}
+
+	return nil
+}
+
+func (tc *testContext) containerCommandShouldHaveTrimmedWhitespace() error {
+	if tc.currentJob == nil {
+		return fmt.Errorf("no current Job")
+	}
+
+	if len(tc.currentJob.Spec.Template.Spec.Containers) == 0 {
+		return fmt.Errorf("Job %s has no containers", tc.currentJob.Name)
+	}
+
+	container := tc.currentJob.Spec.Template.Spec.Containers[0]
+	for _, cmd := range container.Command {
+		if strings.TrimSpace(cmd) != cmd {
+			return fmt.Errorf("Container %s command element has untrimmed whitespace: %q", container.Name, cmd)
+		}
+	}
+
+	return nil
+}
+
+func (tc *testContext) containerShouldHaveProviderEnvVars() error {
+	// Validates that provider environment variables are present
+	// This is a general check that the container has env vars
+	if tc.currentJob == nil {
+		return fmt.Errorf("no current Job")
+	}
+
+	if len(tc.currentJob.Spec.Template.Spec.Containers) == 0 {
+		return fmt.Errorf("Job %s has no containers", tc.currentJob.Name)
+	}
+
+	// Just verify that there are env vars
+	container := tc.currentJob.Spec.Template.Spec.Containers[0]
+	if len(container.Env) == 0 {
+		return fmt.Errorf("Container %s has no environment variables", container.Name)
+	}
+
+	return nil
+}
